remove redundant component
